Create README.md
Dinesh Industries” established in the year 2004. Has made a formidable name for itself as a manufacturer of pressure gauge accessories and components .Our products are sought after in different parts of the world and are known for efficiency, durability, precision engineering and accurate dimensions. Our employees are cornerstones of maturity, innovation and creativity and have brought their talents for the benefit of the organization. They help in building our knowledgeable and capable employees on a path of success by helping them provide quality products to the customers. Great services and usage of recent technological knowledge are the trademarks of the organization all products are made using high quality raw material and are attested by the certified quality management system of the firm. Our products used in all industrial sector and used in pressure gauge.